+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/dmd/DMDDeviceIniService.java
@@ -47,10 +47,15 @@ public class DMDDeviceIniService {
   @Autowired
   private SystemService systemService;
 
+  @Nullable
   public DMDDeviceIniConfiguration getDmdDeviceIni(@NonNull GameEmulator gameEmulator) {
     loadDmdDeviceIni(gameEmulator);
 
     INIConfiguration dmdDeviceIni = dmdDeviceIniFiles.get(gameEmulator.getId());
+    if (dmdDeviceIni == null) {
+      LOG.warn("No dmddevice.ini found for {}", gameEmulator.getName());
+      return null;
+    }
 
     DMDDeviceIniConfiguration config = new DMDDeviceIniConfiguration();
     config.setEmulatorId(gameEmulator.getId());
@@ -271,7 +276,6 @@ public class DMDDeviceIniService {
 
   public String getStoreName(Game game) {
     String storeName = StringUtils.defaultString(game.getRomAlias(), game.getRom());
-    ;
     if (DMDPackageTypes.UltraDMD.equals(game.getDMDType())) {
       storeName = FilenameUtils.getBaseName(game.getGameFileName());
       // cf https://github.com/vbousquet/flexdmd/blob/6357c1874e896777a53348094eafa86f386dd8fe/FlexDMD/FlexDMD.cs#L188
@@ -282,6 +286,7 @@ public class DMDDeviceIniService {
       // cf https://github.com/vbousquet/flexdmd/blob/6357c1874e896777a53348094eafa86f386dd8fe/FlexDMD/FlexDMD.cs#L188
       storeName = storeName.replaceAll("[\\s_vV][\\d_\\.]+[a-z]?(-DOF)?\\*?$", "").trim();
     }
+    storeName = storeName.replaceAll("\\.", " ");
     return storeName;
   }
 
@@ -320,7 +325,7 @@ public class DMDDeviceIniService {
     GameEmulator emulator = game.getEmulator();
     if (emulator != null) {
       DMDDeviceIniConfiguration dmdDeviceIni = getDmdDeviceIni(emulator);
-      if (!dmdDeviceIni.isUseRegistry()) {
+      if (dmdDeviceIni != null && !dmdDeviceIni.isUseRegistry()) {
         INIConfiguration iniConfiguration = getIniConfiguration(game);
         if (iniConfiguration != null) {
           String dmdStoreName = getStoreName(game);